fix compilation failure in CI
