+++ hibernate-core/src/test/java/org/hibernate/orm/test/query/named/BookRepository.java
@@ -7,5 +7,7 @@ package org.hibernate.orm.test.query.named;
 import jakarta.data.repository.Repository;
 
 @Repository
-public interface BookRepository extends BookRepositoryQueries, GenericBookRepository<Jpa4StaticQueryRegistrationTest.Book> {
+public interface BookRepository
+		extends BookRepositoryQueries,
+				GenericBookRepository<Book> {
 }
