+++ hibernate-core/src/test/java/org/hibernate/orm/test/query/named/BookRepositoryQueries.java
@@ -6,9 +6,10 @@ package org.hibernate.orm.test.query.named;
 
 import java.util.List;
 
+import jakarta.data.repository.Param;
 import jakarta.persistence.query.JakartaQuery;
 
 public interface BookRepositoryQueries {
 	@JakartaQuery( "from Jpa4StaticQueryBook where title = :title" )
-	List<Jpa4StaticQueryRegistrationTest.Book> inheritedFindByTitle(String title);
+	List<Book> inheritedFindByTitle(@Param("title") String title);
 }
